ref: redis options
